+++ fsl_mrs/tests/testdata/dynamic/simple_linear_model.py
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:1fe2eae519a3cc706ab3526fa243d3b9324dfd58c51329e1e6a5c8c13454e4f4
-size 991
+oid sha256:92551cf8c71d29bb6ccd03421bdacde9ba32433e29092cdcc5958ab9232bc6d2
+size 962